0.7: EARTH1_PRECISION execution modes + equivalence study instruments
earth1/precision.py: apply_precision walks the World reflectively
(aliasing-preserving; sparse converted in place), float32 and the
pre-registered float16-control degradation (quantize-at-load, run in
f32). Producers that reassign f64 mid-tick are folded back once per
day boundary by recoerce() — an attribute check (no-op) for the f64
reference. float64_survivors() proves zero leaks over 15 f32 days.

Dtype-following edits (all bit-exact no-ops in f64, pinned hash
unchanged): plasticity working copies/rewire delta, propagate pole/X,
conviction pole, spread staging, _add_mutual and household deltas,
knowledge/flourishing adj matvec operands.

earth1/observables.py: protocol families 1-13 collector.
scripts/precision_ab.py: the frozen quadruple study runner (8 pairs
x f64/f32 + 3-pair f16 control, days 3/15/30, precision stamped in
every member record). Full suite 1079 green.

+++ earth1/flourishing.py
@@ -204,7 +204,7 @@ def flourishing_tick(civ, life, fl: Flourishing, kn, health, rng,
 
     # MEANING: art made and art received, plus work worth doing
     made = (kn.works_made > 0)
-    art_flow = np.asarray(civ.adj @ made.astype(np.float64)).ravel() / deg
+    art_flow = np.asarray(civ.adj @ made.astype(civ.adj.dtype)).ravel() / deg
     fl.art_received += art_flow * dt_days * 0.01
     mean_target = np.clip(0.25 + 0.35 * np.clip(fl.art_received, 0, 1)
                           + 0.20 * kn.status + 0.20 * fl.belonging, 0, 1)
